Move influence base class to new module
