+++ src/pydvl/influence/__init__.py
@@ -5,6 +5,6 @@ This package contains algorithms for the computation of the influence function.
 In particular, the package structure and basic API will probably change.
 
 """
+from .base_influence_model import InfluenceType
 from .general import compute_influence_factors, compute_influences
 from .inversion import InversionMethod
-from .twice_differentiable import InfluenceType
